Add PayStatusView and simplify ManagementView
